refactor: refactor rendering
separate object & hud render phases
