+++ src/client/render/scene.js
@@ -5,6 +5,45 @@ import renderCrosshair from './crosshair';
 import renderSpace from './space';
 import renderMap from './map';
 import renderScores from './scores';
+import renderExplosion from './explosion';
+
+function renderObjects(ctx, scene, particles, viewport) {
+  ctx.save();
+  ctx.scale(1, -1);
+  ctx.rotate(-viewport.orientation);
+  ctx.scale(viewport.scale, viewport.scale);
+  ctx.translate(-viewport.position.x, -viewport.position.y);
+
+  renderSpace(ctx, viewport, particles);
+  for (const ship of scene.ships) {
+    renderShip(ctx, ship, viewport);
+  }
+  for (const projectile of scene.projectiles) {
+    renderProjectile(ctx, projectile);
+  }
+  for (const explosion of scene.explosions) {
+    renderExplosion(ctx, scene, explosion);
+  }
+
+  ctx.restore();
+}
+
+function renderHUD(ctx, scene, viewport) {
+  ctx.save();
+  if (!viewport.alive) {
+    ctx.save();
+    ctx.font = '72px Helvetica, sans-serif';
+    ctx.textAlign = 'center';
+    ctx.baseLine = 'middle';
+    ctx.fillText('You\'re dead!  ͡° ͜ʖ ͡°', 0, 0);
+    ctx.restore();
+  } else {
+    renderCrosshair(ctx, viewport);
+    renderMap(ctx, scene, viewport);
+  }
+  renderScores(ctx, scene);
+  ctx.restore();
+}
 
 export default (ctx, scene, particles, viewport) => {
   ctx.save();
@@ -15,47 +54,8 @@ export default (ctx, scene, particles, viewport) => {
     ctx.restore();
     return;
   }
-  ctx.scale(1, -1);
-  ctx.rotate(-viewport.orientation);
-  ctx.scale(viewport.scale, viewport.scale);
-  ctx.translate(-viewport.position.x, -viewport.position.y);
 
-  renderSpace(ctx, viewport, particles);
-
-  for (const ship of scene.ships) {
-    renderShip(ctx, ship, viewport);
-  }
-
-  for (const projectile of scene.projectiles) {
-    renderProjectile(ctx, projectile);
-  }
-
-  for (const explosion of scene.explosions) {
-    if (scene.time > explosion.time + 5000) {
-      continue;
-    }
-    const progress = (scene.time - explosion.time) / 5000;
-    ctx.save();
-    ctx.fillStyle = `rgba(255, 0, 0, ${0.5 - progress / 2})`;
-    ctx.translate(explosion.position.x, explosion.position.y);
-    ctx.beginPath();
-    ctx.arc(0, 0, progress * 100, 0, Math.PI * 2);
-    ctx.fill();
-    ctx.restore();
-  }
-
-  if (!viewport.alive) {
-    ctx.save();
-    ctx.font = '72px Helvetica, sans-serif';
-    ctx.textAlign = 'center';
-    ctx.baseLine = 'middle';
-    ctx.scale(1, -1);
-    ctx.fillText('You\'re dead!  ͡° ͜ʖ ͡°', 0, 0);
-    ctx.restore();
-  } else {
-    renderCrosshair(ctx, viewport);
-    renderMap(ctx, scene, viewport);
-  }
-  renderScores(ctx, scene, viewport);
+  renderObjects(ctx, scene, particles, viewport);
+  renderHUD(ctx, scene, viewport);
   ctx.restore();
 };
